Split follow and object tests
