+++ src/middleware/tests/activitypub/initialize.js
@@ -1,0 +1,33 @@
+const { TripleStoreService } = require('@semapps/triplestore');
+const { LdpService, getPrefixJSON } = require('@semapps/ldp');
+const { ActivityPubService } = require('@semapps/activitypub');
+const CONFIG = require('../config');
+const ontologies = require('../ontologies');
+
+const initialize = broker => async () => {
+  broker.createService(TripleStoreService, {
+    settings: {
+      sparqlEndpoint: CONFIG.SPARQL_ENDPOINT,
+      mainDataset: CONFIG.MAIN_DATASET,
+      jenaUser: CONFIG.JENA_USER,
+      jenaPassword: CONFIG.JENA_PASSWORD
+    }
+  });
+  broker.createService(LdpService, {
+    settings: {
+      baseUrl: CONFIG.HOME_URL + 'ldp/',
+      ontologies
+    }
+  });
+  broker.createService(ActivityPubService, {
+    settings: {
+      baseUri: CONFIG.HOME_URL,
+      additionalContext: getPrefixJSON(ontologies)
+    }
+  });
+
+  await broker.start();
+  await broker.call('triplestore.dropAll');
+};
+
+module.exports = initialize;
